--- backend/homework/homework-3/src/main/java/kdu/ques1/Nurse.java
@@ -1,13 +1,0 @@
-package kdu.ques1;
-
-public class Nurse extends Staff {
-    private long nurseId;
-
-    public long getNurseId() {
-        return nurseId;
-    }
-
-    public void setNurseId(long nurseId) {
-        this.nurseId = nurseId;
-    }
-}
