--- backend/homework/homework-3/src/main/java/kdu/ques1/Patient.java
@@ -1,23 +1,0 @@
-package kdu.ques1;
-
-
-public class Patient extends User {
-    private long patientId;
-    private boolean insured;
-
-    public long getPatientId() {
-        return patientId;
-    }
-
-    public void setPatientId(long patientId) {
-        this.patientId = patientId;
-    }
-
-    public boolean isInsured() {
-        return insured;
-    }
-
-    public void setInsured(boolean insured) {
-        this.insured = insured;
-    }
-}
