--- backend/homework/homework-3/src/main/java/kdu/ques2/InsurencePlan/GoldPlan.java
@@ -1,10 +1,0 @@
-package kdu.ques2.InsurencePlan;
-
-import kdu.ques2.HealthInsurancePlan;
-
-public class GoldPlan extends HealthInsurancePlan {
-    public GoldPlan(){
-        setCoverage(0.8);
-        setDiscount(40.0);
-    }
-}
