--- backend/homework/homework-3/src/main/java/kdu/ques2/Patient.java
@@ -1,32 +1,0 @@
-package kdu.ques2;
-
-public class Patient {
-    private long patientId;
-    private boolean insured;
-
-    public long getPatientId() {
-        return patientId;
-    }
-
-    public void setPatientId(long patientId) {
-        this.patientId = patientId;
-    }
-
-    public boolean isInsured() {
-        return insured;
-    }
-
-    public void setInsured(boolean insured) {
-        this.insured = insured;
-    }
-
-    private HealthInsurancePlan insurancePlan;
-
-    public HealthInsurancePlan getInsurancePlan() {
-        return insurancePlan;
-    }
-
-    public void setInsurancePlan(HealthInsurancePlan insurancePlan) {
-        this.insurancePlan = insurancePlan;
-    }
-}
