--- backend/homework/homework-3/src/main/java/kdu/ques3/InsurencePlan/SilverPlan.java
@@ -1,15 +1,0 @@
-package kdu.ques3.InsurencePlan;
-
-import kdu.ques3.HealthInsurancePlan;
-
-public class SilverPlan extends HealthInsurancePlan {
-    public SilverPlan(){
-        setCoverage(0.7);
-        setDiscount(30.0);
-    }
-
-    @Override
-    public double computeMonthlyPremium(double salary) {
-        return 0.06 * salary;
-    }
-}
