--- backend/homework/homework-3/src/main/java/kdu/ques3/User.java
@@ -1,74 +1,0 @@
-package kdu.ques3;
-
-import kdu.ques2.HealthInsurancePlan;
-
-public class User {
-    private long id;
-    // add rest of the variables
-    private String firstName;
-
-    private String lastName;
-    private String email;
-    private String gender;
-
-    private boolean insured;
-
-    private HealthInsurancePlan insurancePlan;
-
-    public HealthInsurancePlan getInsurancePlan() {
-        return insurancePlan;
-    }
-
-    public void setInsurancePlan(HealthInsurancePlan insurancePlan) {
-        this.insurancePlan = insurancePlan;
-    }
-
-    public boolean isInsured() {
-        return insured;
-    }
-
-    public void setInsured(boolean insured) {
-        this.insured = insured;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getGender() {
-        return gender;
-    }
-
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
-}
-
