--- backend/homework/homework-3/src/main/java/kdu/ques4/HealthInsurancePlan.java
@@ -1,34 +1,0 @@
-package kdu.ques4;
-
-public abstract class HealthInsurancePlan {
-    private InsuranceBrand offeredBy;
-
-    public InsuranceBrand getOfferedBy() {
-        return offeredBy;
-    }
-
-    public void setOfferedBy(InsuranceBrand offeredBy) {
-        this.offeredBy = offeredBy;
-    }
-
-    public abstract double computeMonthlyPremium(double salary, int age, boolean smoking);
-    private double coverage;
-    private double discount;
-
-    public double getDiscount() {
-        return discount;
-    }
-
-    public void setDiscount(double discount) {
-        this.discount = discount;
-    }
-
-    public double getCoverage() {
-        return coverage;
-    }
-
-    public void setCoverage(double coverage) {
-        this.coverage = coverage;
-    }
-
-}
