+++ backend/homework/homework-3/src/main/java/kdu/quesfour/InsuranceBrand.java
@@ -1,0 +1,5 @@
+package kdu.quesfour;
+
+public interface InsuranceBrand {
+    public double computeMonthlyPremium(HealthInsurancePlan insurancePlan, int age, boolean smoking);
+}
