+++ backend/homework/homework-3/src/main/java/kdu/quesfour/InsurencePlan/GoldPlan.java
@@ -1,0 +1,14 @@
+package kdu.quesfour.InsurencePlan;
+
+import kdu.quesfour.HealthInsurancePlan;
+
+public class GoldPlan extends HealthInsurancePlan {
+    public GoldPlan(){
+        setCoverage(0.8);
+        setDiscount(40.0);
+    }
+    @Override
+    public double computeMonthlyPremium(double salary, int age, boolean smoking) {
+        return 0.07 * salary + getOfferedBy().computeMonthlyPremium(this, age, smoking);
+    }
+}
