+++ backend/homework/homework-3/src/main/java/kdu/quesone/Nurse.java
@@ -1,0 +1,13 @@
+package kdu.quesone;
+
+public class Nurse extends Staff {
+    private long nurseId;
+
+    public long getNurseId() {
+        return nurseId;
+    }
+
+    public void setNurseId(long nurseId) {
+        this.nurseId = nurseId;
+    }
+}
