+++ backend/homework/homework-3/src/main/java/kdu/questhree/InsurencePlan/BronzePlan.java
@@ -1,0 +1,15 @@
+package kdu.questhree.InsurencePlan;
+
+import kdu.questhree.HealthInsurancePlan;
+
+public class BronzePlan extends HealthInsurancePlan {
+    public BronzePlan(){
+        setCoverage(0.6);
+        setDiscount(25.0);
+    }
+
+    @Override
+    public double computeMonthlyPremium(double salary) {
+        return 0.05 * salary;
+    }
+}
