+++ backend/homework/homework-3/src/main/java/kdu/questhree/Patient.java
@@ -1,0 +1,16 @@
+package kdu.questhree;
+
+public class Patient extends User {
+    private long patientId;
+
+    public long getPatientId() {
+        return patientId;
+    }
+
+    public void setPatientId(long patientId) {
+        this.patientId = patientId;
+    }
+
+
+
+}
