+++ backend/homework/homework-3/src/main/java/kdu/questwo/HealthInsurancePlan.java
@@ -1,0 +1,24 @@
+package kdu.questwo;
+
+public class HealthInsurancePlan {
+
+    private double coverage;
+    private double discount;
+
+    public double getDiscount() {
+        return discount;
+    }
+
+    public void setDiscount(double discount) {
+        this.discount = discount;
+    }
+
+    public double getCoverage() {
+        return coverage;
+    }
+
+    public void setCoverage(double coverage) {
+        this.coverage = coverage;
+    }
+
+}
