+++ backend/homework/homework-3/src/main/java/kdu/questwo/InsurencePlan/BronzePlan.java
@@ -1,0 +1,10 @@
+package kdu.questwo.InsurencePlan;
+
+import kdu.questwo.HealthInsurancePlan;
+
+public class BronzePlan extends HealthInsurancePlan {
+    public BronzePlan(){
+        setCoverage(0.6);
+        setDiscount(25.0);
+    }
+}
