+++ backend/homework/homework-3/src/main/java/kdu/questwo/InsurencePlan/GoldPlan.java
@@ -1,0 +1,10 @@
+package kdu.questwo.InsurencePlan;
+
+import kdu.questwo.HealthInsurancePlan;
+
+public class GoldPlan extends HealthInsurancePlan {
+    public GoldPlan(){
+        setCoverage(0.8);
+        setDiscount(40.0);
+    }
+}
