+++ backend/homework/homework-3/src/main/java/kdu/questwo/InsurencePlan/PlatinumPlan.java
@@ -1,0 +1,10 @@
+package kdu.questwo.InsurencePlan;
+
+import kdu.questwo.HealthInsurancePlan;
+
+public class PlatinumPlan extends HealthInsurancePlan {
+    public PlatinumPlan(){
+        setCoverage(0.9);
+        setDiscount(50.0);
+    }
+}
