+++ backend/homework/homework-3/src/main/java/kdu/questwo/InsurencePlan/SilverPlan.java
@@ -1,0 +1,10 @@
+package kdu.questwo.InsurencePlan;
+
+import kdu.questwo.HealthInsurancePlan;
+
+public class SilverPlan extends HealthInsurancePlan {
+    public SilverPlan(){
+        setCoverage(0.7);
+        setDiscount(30.0);
+    }
+}
